backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers


